Towards a generalization of input object in powerstate
