+++ src/mitim_modules/powertorch/scripts/calculateTargets.py
@@ -112,7 +112,7 @@ def calculator(
 
     p.profiles.deriveQuantities()
     
-    p.to_gacode(
+    p.from_powerstate(
         write_input_gacode=folder / "input.gacode.new.powerstate",
         position_in_powerstate_batch=0,
         postprocess_input_gacode={
